+++ src/com/ansorgit/plugins/bash/lang/lexer/_BashLexerBase.java
@@ -1679,11 +1679,23 @@ goToState(S_STRINGMODE); return STRING_BEGIN;
             }
           case 267: break;
           case 82: 
-            { if (heredocState().isNextMarker(yytext())) {
+            { //support end marker followed by a backtick if nested in a backtick command
+            CharSequence markerText = yytext();
+            boolean dropLastChar = false;
+            if (isInState(S_BACKQUOTE) && yylength() >= 2 && yycharat(yylength()-1) == '`') {
+                markerText = markerText.subSequence(0, yylength()-1);
+                dropLastChar = true;
+            }
+
+            if (heredocState().isNextMarker(markerText)) {
                 boolean ignoreTabs = heredocState().isIgnoringTabs();
 
-                heredocState().popMarker(yytext());
+                heredocState().popMarker(markerText);
                 popStates(S_HEREDOC);
+
+                if (dropLastChar) {
+                    yypushback(1);
+                }
 
                 return ignoreTabs ? HEREDOC_MARKER_IGNORING_TABS_END : HEREDOC_MARKER_END;
             }